Added some useful functions, and rearranged.
Added functions for getting text inputs from Post and Query. 
getPost and getQuery do return a 'True' boolean, even if theres no value in the post/query, just as long as the post/query is there.
getInputPost and getInputQuery do return the value of a post/query is there's any, otherwise returns false.